Implemented HandleCustExit, fixed int division... I think

+++ src/main/java/sm/rental/model/outputs/SSOV.java
@@ -18,11 +18,11 @@ public class SSOV
 	public void addSatisfiedCust(){
 		numSatisfied++;
 		numServed++;
-		customerSatisfactionRate = numSatisfied/numServed;
+		customerSatisfactionRate = (double)numSatisfied/(double)numServed;
 	}
 
 	public void addUnsatisfiedCust(){
 		numServed++;
-		customerSatisfactionRate = numSatisfied/numServed;
+		customerSatisfactionRate = (double)numSatisfied/(double)numServed;
 	}
 }
